+++ packages/plugin-network-breadcrumbs/network-breadcrumbs.js
@@ -1,10 +1,5 @@
 const BREADCRUMB_TYPE = 'request'
 
-// keys to safely store metadata on the request object
-const REQUEST_SETUP_KEY = 'BS~~S'
-const REQUEST_URL_KEY = 'BS~~U'
-const REQUEST_METHOD_KEY = 'BS~~M'
-
 const includes = require('@bugsnag/core/lib/es-utils/includes')
 
 /*
@@ -31,23 +26,24 @@ module.exports = (_ignoredUrls = [], win = window) => {
 
         // override native open()
         win.XMLHttpRequest.prototype.open = function open (method, url) {
-          // store url and HTTP method for later
-          this[REQUEST_URL_KEY] = url
-          this[REQUEST_METHOD_KEY] = method
+          let requestSetupKey = false
+
+          const error = () => handleXHRError(method, url)
+          const load = () => handleXHRLoad(method, url, this.status)
 
           // if we have already setup listeners, it means open() was called twice, we need to remove
           // the listeners and recreate them
-          if (this[REQUEST_SETUP_KEY]) {
-            this.removeEventListener('load', handleXHRLoad)
-            this.removeEventListener('error', handleXHRError)
+          if (requestSetupKey) {
+            this.removeEventListener('load', load)
+            this.removeEventListener('error', error)
           }
 
           // attach load event listener
-          this.addEventListener('load', handleXHRLoad)
+          this.addEventListener('load', load)
           // attach error event listener
-          this.addEventListener('error', handleXHRError)
+          this.addEventListener('error', error)
 
-          this[REQUEST_SETUP_KEY] = true
+          requestSetupKey = true
 
           nativeOpen.apply(this, arguments)
         }
@@ -59,9 +55,7 @@ module.exports = (_ignoredUrls = [], win = window) => {
         }
       }
 
-      function handleXHRLoad () {
-        const url = this[REQUEST_URL_KEY]
-
+      function handleXHRLoad (method, url, status) {
         if (url === undefined) {
           client._logger.warn('The request URL is no longer present on this XMLHttpRequest. A breadcrumb cannot be left for this request.')
           return
@@ -74,10 +68,10 @@ module.exports = (_ignoredUrls = [], win = window) => {
           return
         }
         const metadata = {
-          status: this.status,
-          request: `${this[REQUEST_METHOD_KEY]} ${this[REQUEST_URL_KEY]}`
+          status: status,
+          request: `${method} ${url}`
         }
-        if (this.status >= 400) {
+        if (status >= 400) {
           // contacted server but got an error response
           client.leaveBreadcrumb('XMLHttpRequest failed', metadata, BREADCRUMB_TYPE)
         } else {
@@ -85,9 +79,7 @@ module.exports = (_ignoredUrls = [], win = window) => {
         }
       }
 
-      function handleXHRError () {
-        const url = this[REQUEST_URL_KEY]
-
+      function handleXHRError (method, url) {
         if (url === undefined) {
           client._logger.warn('The request URL is no longer present on this XMLHttpRequest. A breadcrumb cannot be left for this request.')
           return
@@ -100,7 +92,7 @@ module.exports = (_ignoredUrls = [], win = window) => {
 
         // failed to contact server
         client.leaveBreadcrumb('XMLHttpRequest error', {
-          request: `${this[REQUEST_METHOD_KEY]} ${this[REQUEST_URL_KEY]}`
+          request: `${method} ${url}`
         }, BREADCRUMB_TYPE)
       }